[feat] Implement JWT auth

+++ app/Actions/Task/DestroyTaskAction.php
@@ -2,7 +2,9 @@
 
 namespace App\Actions\Task;
 
+use App\Models\Todo;
 use App\Repositories\Interfaces\TaskRepositoryInterface;
+use Illuminate\Support\Facades\Auth;
 
 class DestroyTaskAction
 {
@@ -12,6 +14,12 @@ class DestroyTaskAction
 
     public function execute(int $id): void
     {
+        $taskWithUser = $this->taskRepository->findUserAndTask($id);
+
+        if (Auth::user()->cannot('check', [Todo::class, $taskWithUser->todo->user_id])) {
+            abort(403);
+        }
+
         $this->taskRepository->destroy($id);
     }
 }
